Use a mapperScript instead of a separate class
This is the most sensible default I could think of -- it matches the
django-lti-launch user id of smashing together the first 14 characters
of the tool consumer instance guid, and the first 14 characters of
the LTI user_id.  We will have to validate that this is giving us
a unique enough user ID somehow, or just give up and use a hash of
the whole value, or something.

We'll have to override the "LTI" usermapper anyway in our project
code, to implement the extra ACL and group membership logic for the
specific access control requirements.