Add ltiRoles to ltiuser.xsd
User document models now contain an "ltiRoles" property, which is
a direct copy of the "roles" LTI launch parameter.

I tried to use the `nxs:stringList` parameter type to better match
the appropriate usage of the data (LTI specifies it to be a comma-
separated string) but the test suite did not seem to correctly
create the `users` table.  I decided this was not a windmill worth
tilting at right now.